@@ -1,6 +1,6 @@
-# codebase will be a class that gets created from an active websocket connection
+# Codebase will be a class that gets created from an active websocket connection
 
-# methods called on the codebase will be converted into WS requests send to the client
+# methods called on the Codebase will be converted into WS requests send to the client
 # then a listener will be created to listen for the response and then return the response to the caller
 
-# codebase is a class that also exsists on the client side, it will be used to execute the requests from the server
+# Codebase is a class that also exsists on the client side, it will be used to execute the requests from the server
