@@ -2,29 +2,34 @@ import asyncio
 from typing import AsyncIterator
 
 from funcchain.schema.types import UniversalChatModel as LLM
-from shared.codebase_interface import CodebaseInterface
+from shared.codebase.core import Codebase
+from shared.schemas import CodebaseIOResult, Data
 from websockets import WebSocketClientProtocol
 
-from .shared.schemas import CodeBaseIOResult
+# todo have a local running script that checks all incoming shell commands for security reasons
+# let the user see set his own model optional local model
 
 
-class Codebase(CodebaseInterface):
+class CodebaseClient(Codebase):
     def __init__(self, websocket: WebSocketClientProtocol, llm: LLM) -> None:
         self.websocket = websocket
         self.llm: LLM = llm
 
-    async def stream_shell(self, cmd: str) -> AsyncIterator[CodeBaseIOResult]:  # type: ignore
+    async def shell(self, cmd: str) -> str:
+        return "todo"
+
+    async def stream_shell(self, cmd: str) -> AsyncIterator[CodebaseIOResult]:  # type: ignore
         process = await asyncio.create_subprocess_shell(
             cmd, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
         )
         while True:
             if process.stdout and (output := await process.stdout.readline()):
-                yield CodeBaseIOResult(
+                yield CodebaseIOResult(
                     data=output.decode(),
                     status="healthy",
                 )
             if process.stderr and (error := await process.stderr.readline()):
-                yield CodeBaseIOResult(
+                yield CodebaseIOResult(
                     data=error.decode(),
                     status="error",
                 )
@@ -32,4 +37,17 @@ class Codebase(CodebaseInterface):
                 break
 
         await process.wait()
-        yield CodeBaseIOResult(data="")
+        yield CodebaseIOResult(data="")
+
+    async def read_file(self, path: str) -> CodebaseIOResult:
+        with open(path, "r") as file:
+            return CodebaseIOResult(data=file.read(), status="healthy")
+
+    async def write_file(self, path: str, data: Data) -> CodebaseIOResult:
+        if isinstance(data, bytes):
+            data = data.decode()
+
+        with open(path, "w") as file:
+            file.write(data)
+
+        return CodebaseIOResult(data="success", status="healthy")
