@@ -6,30 +6,30 @@ from typing import Any, Coroutine, Union
 
 import yaml  # type: ignore
 
-from .file import CodeBaseFile
+from .file import CodebaseFile
 from .ignore import is_ignored_by_gitignore
-from .node import CodeBaseNode
+from .node import CodebaseNode
 
 
-class CodeBaseTree(CodeBaseNode):
-    nodes: list[CodeBaseNode] = []
+class CodebaseTree(CodebaseNode):
+    nodes: list[CodebaseNode] = []
 
     def __init__(self, path: Path | str, **data: Any) -> None:
         data["name"] = Path(path).as_posix()
         super().__init__(**data)
 
     @classmethod
-    def from_dict(cls, data: dict) -> "CodeBaseTree":
+    def from_dict(cls, data: dict) -> "CodebaseTree":
         path = data.pop("name")
         nodes_data = data.pop("nodes")
         nodes = [
-            CodeBaseFile.from_dict(node_data) if "summary" in node_data else CodeBaseTree.from_dict(node_data)
+            CodebaseFile.from_dict(node_data) if "summary" in node_data else CodebaseTree.from_dict(node_data)
             for node_data in nodes_data
         ]
         return cls(path, nodes=nodes, **data)
 
     @classmethod
-    async def load(cls) -> "CodeBaseTree":
+    async def load(cls) -> "CodebaseTree":
         if Path(".context/tree.yaml").exists():
             with open(".context/tree.yaml", "r") as f:
                 data = yaml.safe_load(f)
@@ -41,15 +41,15 @@ class CodeBaseTree(CodeBaseNode):
             return await cls.new()
 
     @classmethod
-    async def from_path(cls, path: Path) -> "CodeBaseTree":
-        tasks: list[Coroutine[Any, Any, CodeBaseNode]] = [
-            CodeBaseFile.from_path(file_path) if file_path.is_file() else cls.from_path(file_path)
+    async def from_path(cls, path: Path) -> "CodebaseTree":
+        tasks: list[Coroutine[Any, Any, CodebaseNode]] = [
+            CodebaseFile.from_path(file_path) if file_path.is_file() else cls.from_path(file_path)
             for file_path in path.iterdir()
             if not is_ignored_by_gitignore(file_path.as_posix())
         ]
         if len(tasks) > 50:
             input(f"Found {len(tasks)} files in {path}. Press enter to continue...")
-        nodes: list[CodeBaseNode] = await asyncio.gather(*tasks)
+        nodes: list[CodebaseNode] = await asyncio.gather(*tasks)
 
         folder_hash = hashlib.sha256(("".join(str(node.sha256) for node in nodes)).encode()).hexdigest()
 
@@ -62,14 +62,14 @@ class CodeBaseTree(CodeBaseNode):
         return tree
 
     @classmethod
-    async def new(cls) -> "CodeBaseTree":
+    async def new(cls) -> "CodebaseTree":
         return await cls.from_path(Path("."))
 
-    async def refresh(self) -> "CodeBaseTree":
-        # gather new nodes from codebase not in self.nodes
+    async def refresh(self) -> "CodebaseTree":
+        # gather new nodes from Codebase not in self.nodes
         node_paths = [node.path for node in self.nodes]
         new_node_tasks = [
-            CodeBaseFile.from_path(file_path) if file_path.is_file() else CodeBaseTree.from_path(file_path)
+            CodebaseFile.from_path(file_path) if file_path.is_file() else CodebaseTree.from_path(file_path)
             for file_path in self.path.iterdir()
             if not is_ignored_by_gitignore(file_path.as_posix()) and file_path not in node_paths
         ]
@@ -81,7 +81,7 @@ class CodeBaseTree(CodeBaseNode):
             in [file_path for file_path in self.path.iterdir() if not is_ignored_by_gitignore(file_path.as_posix())]
         ]
 
-        # delete nodes not in codebase anymore
+        # delete nodes not in Codebase anymore
         ignored_nodes = [
             file_path for file_path in self.path.iterdir() if not is_ignored_by_gitignore(file_path.as_posix())
         ]
@@ -102,13 +102,13 @@ class CodeBaseTree(CodeBaseNode):
         return self
 
     @property
-    def files(self) -> list[CodeBaseFile]:
+    def files(self) -> list[CodebaseFile]:
         files = []
         for node in self.nodes:
-            if isinstance(node, CodeBaseFile):
+            if isinstance(node, CodebaseFile):
                 files.append(node)
             else:
-                assert isinstance(node, CodeBaseTree)
+                assert isinstance(node, CodebaseTree)
                 files.extend(node.files)
         return files
 
@@ -120,7 +120,7 @@ class CodeBaseTree(CodeBaseNode):
         return folder_str
 
     def __repr__(self) -> str:
-        return f"CodeBaseTree(path={self.path}, files={len(self.files)}, nodes={len(self.nodes)})"
+        return f"CodebaseTree(path={self.path}, files={len(self.files)}, nodes={len(self.nodes)})"
 
     def show(self) -> str:
         return self.__str__()
