@@ -6,7 +6,7 @@ Data = str | bytes
 
 
 class WSMessage(BaseModel):
-    type: Literal["msg", "error", "action", "codebaseio"] = Field(..., description="Type of the object.")
+    type: Literal["msg", "error", "action", "Codebaseio"] = Field(..., description="Type of the object.")
     data: BaseModel | str = Field(..., description="Data of the object.")
     action: str | None = Field(None, description="Action class name.")
 
@@ -26,6 +26,6 @@ class ResultModel(BaseModel):
 class ActionResult(ResultModel): ...
 
 
-class CodeBaseIOResult(ResultModel):
+class CodebaseIOResult(ResultModel):
     data: Data | None = None
     path: str | None = None
