@@ -1,24 +1,27 @@
 import asyncio
 import os
-from asyncio import subprocess
+from asyncio import subprocess as asp
 
 
 async def prepare_environments(example_path: str) -> None:
-    await subprocess.create_subprocess_shell(
-        "python3 -m venv .venv && source venv/bin/activate &&pip install -r requirements.txt",
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        cwd="playgrounds/" + example_path,
+    await asp.create_subprocess_shell(
+        # todo optimize with uv
+        "python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt",
+        cwd="tests/codebases/playgrounds/" + example_path,
     )
 
 
 async def run_debugging(example_path: str) -> None:
-    await subprocess.create_subprocess_shell(
-        "conda activate neural-cli && codr debug 'venv/bin/python3 main.py'",
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        cwd="playgrounds/" + example_path,
+    await asp.create_subprocess_shell(
+        "codr debug '.venv/bin/python3 main.py'",
+        stdout=asp.PIPE,
+        stderr=asp.PIPE,
+        cwd="tests/codebases/playgrounds/" + example_path,
     )
+
+
+async def reset_environments(example_path: str) -> None:
+    await asp.create_subprocess_shell("rm -rf playgrounds", cwd="tests/codebases/")
 
 
 async def fix_codebases() -> None:
@@ -33,7 +36,7 @@ async def fix_codebases() -> None:
 
 
 def test_fix_codebases() -> None:
-    assert True
+    asyncio.run(fix_codebases())
 
 
 if __name__ == "__main__":
